Created GlobalStyles files and modified DefaultTheme's interface of Styled-C, installed chromatic and uploaded stories
