+++ .storybook/preview.cjs
@@ -1,5 +1,4 @@
 import { ThemeProvider, withBackground } from './decorators';
-import '../src/index.css';
 import 'bootstrap/dist/css/bootstrap.min.css';
 
 export const parameters = {
