+++ package.json
@@ -8,7 +8,8 @@
     "build": "tsc && vite build",
     "preview": "vite preview",
     "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook"
+    "build-storybook": "build-storybook",
+    "chromatic": "npx chromatic --project-token=6515ea744185"
   },
   "dependencies": {
     "@reduxjs/toolkit": "^1.9.3",
@@ -37,6 +38,7 @@
     "@typescript-eslint/parser": "^5.53.0",
     "@vitejs/plugin-react": "^3.1.0",
     "babel-loader": "^8.3.0",
+    "chromatic": "^6.17.1",
     "eslint": "8.2.0",
     "eslint-config-airbnb": "19.0.4",
     "eslint-config-airbnb-typescript": "^17.0.0",
@@ -53,5 +55,7 @@
   },
   "resolutions": {
     "styled-components": "^5"
-  }
+  },
+  "readme": "ERROR: No README data found!",
+  "_id": "friends_app@0.0.0"
 }
