+++ src/Theme/Theme.tsx
@@ -1,12 +1,11 @@
 import { FC, ReactNode, useReducer, useMemo } from 'react';
-import { ThemeProvider } from 'styled-components';
+import { ThemeProvider, DefaultTheme } from 'styled-components';
+import { GlobalStyles } from './GlobalStyles';
 
 import 'bootstrap/dist/css/bootstrap.min.css';
 
 import { actions, reducer, ThemeContext } from './Context';
 import { initialTheme } from './initial-theme';
-
-import type { Theme } from './types';
 
 interface ThemeProps {
   children: ReactNode;
@@ -18,7 +17,7 @@ export const FriendThemeProvider: FC<ThemeProps> = ({ children }) => {
   const value = useMemo(
     () => ({
       theme: state.theme,
-      updateTheme: (theme: Theme) => {
+      updateTheme: (theme: DefaultTheme) => {
         dispatch({ type: actions.UPDATE_THEME, payload: theme });
       },
     }),
@@ -27,7 +26,10 @@ export const FriendThemeProvider: FC<ThemeProps> = ({ children }) => {
 
   return (
     <ThemeContext.Provider value={value}>
-      <ThemeProvider theme={state.theme}>{children}</ThemeProvider>
+      <ThemeProvider theme={state.theme}>
+        <GlobalStyles />
+        {children}
+      </ThemeProvider>
     </ThemeContext.Provider>
   );
 };
