+++ src/Theme/index.ts
@@ -1,5 +1,6 @@
 export { useThemeContext } from './Context';
 export { FriendThemeProvider } from './Theme';
-export type { FriendTheme, Theme } from './types';
+export type { FriendTheme } from './types';
 export { initialTheme } from './initial-theme';
 export { ThemeContext } from './Context';
+export { GlobalStyles } from './GlobalStyles';
