--- src/index.css
@@ -1,23 +1,0 @@
-:root {
-  --fr-main-gray70: #B1B9DB;
-}
-
-* {
-  margin: 0;
-  padding: 0;
-}
-*,
-*::before,
-*::after {
-  box-sizing: border-box;
-}
-
-html {
-  box-sizing: border-box;
-  scroll-behavior: smooth;
-}
-
-p, h1, h2, h3, h4, h5, a {
-  margin: 0;
-  padding: 0;
-}
